requirements.txt
openai
datetime
python-dateutil
gpiozero
spidev
pillow
python-dotenv
logging